+++ CHANGELOG.md
@@ -2,7 +2,7 @@ git # Deoxys Changelog
 
 ## Next release
 
-- fix: storage nonce
+- fix: storage nonce and key/value
 - fix: class and store updates and block desync after ctrl+c
 - fix: compile without libm
 - fix: genesis state_update